requirements.txt
newspaper3k
pprint
